+++ src/Symfony/Component/Form/Extension/DependencyInjection/DependencyInjectionExtension.php
@@ -58,9 +58,19 @@ class DependencyInjectionExtension implements FormExtensionInterface
             foreach ($this->typeExtensionServices[$name] as $serviceId => $extension) {
                 $extensions[] = $extension;
 
-                // validate result of getExtendedType() to ensure it is consistent with the service definition
-                if ($extension->getExtendedType() !== $name) {
-                    throw new InvalidArgumentException(sprintf('The extended type specified for the service "%s" does not match the actual extended type. Expected "%s", given "%s".', $serviceId, $name, $extension->getExtendedType()));
+                if (method_exists($extension, 'getExtendedTypes')) {
+                    $extendedTypes = array();
+
+                    foreach ($extension::getExtendedTypes() as $extendedType) {
+                        $extendedTypes[] = $extendedType;
+                    }
+                } else {
+                    $extendedTypes = array($extension->getExtendedType());
+                }
+
+                // validate the result of getExtendedTypes()/getExtendedType() to ensure it is consistent with the service definition
+                if (!\in_array($name, $extendedTypes, true)) {
+                    throw new InvalidArgumentException(sprintf('The extended type specified for the service "%s" does not match the actual extended type. Expected "%s", given "%s".', $serviceId, $name, implode(', ', $extendedTypes)));
                 }
             }
         }