+++ src/Symfony/Component/Form/Extension/Validator/Type/SubmitTypeValidatorExtension.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Form\Extension\Validator\Type;
 
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
@@ -19,8 +21,8 @@ class SubmitTypeValidatorExtension extends BaseValidatorExtension
     /**
      * {@inheritdoc}
      */
-    public function getExtendedType()
+    public static function getExtendedTypes(): iterable
     {
-        return 'Symfony\Component\Form\Extension\Core\Type\SubmitType';
+        return array(SubmitType::class);
     }
 }