+++ src/Symfony/Component/Form/Tests/Fixtures/ChoiceTypeExtension.php
@@ -12,17 +12,11 @@
 namespace Symfony\Component\Form\Tests\Fixtures;
 
 use Symfony\Component\Form\AbstractTypeExtension;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class ChoiceTypeExtension extends AbstractTypeExtension
 {
-    private $extendedType;
-
-    public function __construct($extendedType = ChoiceType::class)
-    {
-        $this->extendedType = $extendedType;
-    }
+    public static $extendedType;
 
     /**
      * {@inheritdoc}
@@ -38,8 +32,8 @@ class ChoiceTypeExtension extends AbstractTypeExtension
     /**
      * {@inheritdoc}
      */
-    public function getExtendedType()
+    public static function getExtendedTypes(): iterable
     {
-        return $this->extendedType;
+        return array(self::$extendedType);
     }
 }